Refactor some code (small improvments)
