+++ src/predict/predict.py
@@ -1,14 +1,12 @@
-from pathlib import Path
-from typing import List
-import os
-from datetime import datetime
-import requests
-from typing import Optional
-from custom_logger import logger
 import pandas as pd
 import joblib
+import os
+
+from pathlib import Path
 from sklearn.feature_extraction.text import TfidfVectorizer
-from s3handler import S3Handler #TODO
+
+from src.custom_logger import logger
+from src.s3handler import S3Handler
 
 
 
@@ -23,7 +21,8 @@ def get_csv_file_path(prediction_folder: str):
     """
     Construct the path to the cleaned CSV file based on a folder path.
     """
-    return os.path.join(prediction_folder.rstrip('/'), 'cleaned', 'cleaned.csv')
+    cleaned_dir = get_env_var("PREDICT_CLEANED_DIR")
+    return os.path.join(prediction_folder.rstrip('/'), cleaned_dir, 'cleaned.csv')
 
 def transform_with_tfidf(fitted_vectorizer: TfidfVectorizer, csv_file_path: str) -> pd.DataFrame:
     """
@@ -73,8 +72,11 @@ def make_predictions(cleaned_csv_path, model, vectorized_data, display_predictio
         logger.info("Affichage des prédictions.")
 
     # Définir le chemin de sortie par défaut
-    output_csv_path = Path(cleaned_csv_path).parent.parent / "prediction/predictions.csv"
-    output_json_path = Path(cleaned_csv_path).parent.parent / "prediction/predictions.json"
+    prediction_dir = get_env_var("PREDICT_PREDICTIONS_DIR")
+    output_csv_path = Path(cleaned_csv_path).parent.parent / prediction_dir / "predictions.csv"
+    output_json_path = Path(cleaned_csv_path).parent.parent / prediction_dir / "predictions.json"
+    # output_csv_path = Path(cleaned_csv_path).parent.parent / "prediction/predictions.csv"
+    # output_json_path = Path(cleaned_csv_path).parent.parent / "prediction/predictions.json"
     
     # Création des répertoires si nécessaire
     output_csv_path = Path(output_csv_path)
@@ -100,8 +102,8 @@ def make_predictions(cleaned_csv_path, model, vectorized_data, display_predictio
 def load_vectorizer_and_model():
     """Load the TF-IDF vectorizer and prediction model."""
     logger.info("Loading vectorizer and model ENV variables.")
-    vectorizer_path = get_env_var("PREDICT_TFIDF_VECTORIZER_PATH")
-    model_path = get_env_var("PREDICT_MODEL_PATH")
+    vectorizer_path = get_env_var("DATA_PREPROCESSING_TFIDF_VECTORIZER_PATH")
+    model_path = get_env_var("MODEL_TRAIN_MODEL_TRAIN_PATH")
     print(vectorizer_path, model_path)
     logger.info("Loading vectorizer and model.")
     vectorizer = joblib.load(vectorizer_path)
@@ -148,7 +150,7 @@ def process_predictions(
     logger.info("Predictions generated successfully.")
     
     # Upload prediction results
-    prediction_folder = Path(csv_file_path).parent.parent / "prediction"
+    prediction_folder = Path(csv_file_path).parent.parent / get_env_var("PREDICT_PREDICTIONS_DIR")
     logger.info(f"Uploading predictions to S3: {prediction_folder}")
     s3_handler.upload_directory(
         remote_path=str(prediction_folder), local_directory_name=str(prediction_folder)
@@ -181,26 +183,4 @@ def main(prediction_folder: str):
         raise
 
 if __name__ == "__main__":
-    # tfidf_vectorizer_path = 'models/vectorizers/tfidf_vectorizer.joblib' #TODO get_env_var("DATA_PREPROCESSING_TFIDF_VECTORIZER_PATH")
-    # model_path = 'models/train/ovrc.joblib'# TODO get_env_var("MODEL_TRAIN_MODEL_TRAIN_PATH")
-
-    # fitted_vectorizer =  joblib.load(tfidf_vectorizer_path)
-    # model = joblib.load(model_path)
-
-    # aws_access_key_id = get_env_var("AWS_ACCESS_KEY_ID")
-    # aws_bucket_name = get_env_var("AWS_BUCKET_NAME")
-    # aws_secret_access_key = get_env_var("AWS_SECRET_ACCESS_KEY")
-    # handler = S3Handler(aws_bucket_name)
-
-    # prediction_folder = 'prediction_1731849628.762522/'
-    # csv_file_path = get_csv_file_path(prediction_folder)
-    # csv_file_path = handler.download_file(csv_file_path)
-
-
-    # vectorized_data = transform_with_tfidf(fitted_vectorizer, csv_file_path)
-    # print(make_predictions(csv_file_path, model, vectorized_data))
-
-    # remote_path = local_directory = str(Path(csv_file_path).parent.parent / 'prediction')
-    # handler.upload_directory(remote_path=remote_path, local_directory_name=local_directory)
-
     print(main('prediction_1731849628.762522/'))
